pyproject.toml
[tool.poetry]
name = "homeworckapi"
version = "0.1.0"
description = ""
authors = ["IVadymMarchenko <topim31@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
alembic = "^1.13.1"
fastapi = "^0.110.0"
uvicorn = { extras = ["standard"], version = "^0.29.0" }
sqlalchemy = "^2.0.28"
asyncpg = "^0.29.0"
pydantic = { extras = ["email"], version = "^2.6.4" }
python-jose = { extras = ["cryptography"], version = "^3.3.0" }
passlib = { extras = ["bcrypt"], version = "^1.7.4" }
python-multipart = "^0.0.9"
fastapi-mail = "^1.4.1"
fastapi-limiter = "^0.1.6"
python-dotenv = "^1.0.1"
redis = "^5.0.3"
cloudinary = "^1.39.1"
sphinx-automodapi = "^0.17.0"
pytest = "^8.1.1"


[tool.poetry.group.dev.dependencies]
sphinx = "^7.2.6"


[tool.poetry.group.test.dependencies]
aiosqlite = "^0.20.0"
pytest-asyncio = "^0.23.6"
httpx = "^0.27.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
addopts = "--doctest-modules"
testpath = ["tests"]
pythonpath = "."
filterwarnings='ignore::DeprecationWarning'
